Move functionality from layer store into map store
