+++ web/src/store/map.ts
@@ -1,8 +1,10 @@
 import { defineStore } from 'pinia';
 import { ref, shallowRef, watch } from 'vue';
 import { Map, MapLayerMouseEvent, Popup, Source } from "maplibre-gl";
-import { ClickedFeatureData, Project } from '@/types';
-
+import { ClickedFeatureData, Project, RasterData, RasterDataValues, VectorData, LayerFrame } from '@/types';
+import { getRasterDataValues } from '@/api/rest';
+import { baseURL } from '@/api/auth';
+import proj4 from 'proj4';
 
 export const useMapStore = defineStore('map', () => {
   const map = shallowRef<Map>();
@@ -10,6 +12,7 @@ export const useMapStore = defineStore('map', () => {
   const showMapBaseLayer = ref(true);
   const tooltipOverlay = ref<Popup>();
   const clickedFeature = ref<ClickedFeatureData>();
+  const rasterTooltipDataCache = ref<Record<number, RasterDataValues | undefined>>({});
 
   function handleLayerClick(e: MapLayerMouseEvent) {
     const map = getMap();
@@ -108,6 +111,184 @@ export const useMapStore = defineStore('map', () => {
     });
   }
 
+  function createVectorFeatureMapLayers(source: Source) {
+    const map = getMap();
+    const sourceIdentifiers = source.id.split('.');
+    const metadata = {
+      layer_id: sourceIdentifiers[0],
+      layer_copy_id: sourceIdentifiers[1],
+      frame_id: sourceIdentifiers[2],
+    }
+
+    // Fill Layer
+    map.addLayer({
+      id: source.id + '.fill',
+      type: "fill",
+      source: source.id,
+      metadata,
+      "source-layer": "default",
+      filter: [
+        "match",
+        ["geometry-type"],
+        ["Polygon", "MultiPolygon"],
+        true,
+        false,
+      ],
+      layout: {
+        visibility: "visible",
+      },
+      paint: {
+        "fill-color": "black",
+        "fill-opacity": 0.5,
+      },
+    });
+
+    // Line Layer
+    map.addLayer({
+      id: source.id + '.line',
+      type: "line",
+      source: source.id,
+      metadata,
+      "source-layer": "default",
+      layout: {
+        "line-join": "round",
+        "line-cap": "round",
+        visibility: "visible",
+      },
+      paint: {
+        "line-color": "black",
+        "line-width": 2,
+        "line-opacity": 1,
+      },
+    });
+
+    // Circle Layer
+    map.addLayer({
+      id: source.id + '.circle',
+      type: "circle",
+      source: source.id,
+      metadata,
+      "source-layer": "default",
+      filter: ["match", ["geometry-type"], ["Point", "MultiPoint"], true, false],
+      paint: {
+        "circle-color": "black",
+        "circle-opacity": 1,
+        "circle-stroke-color": "black",
+        "circle-stroke-opacity": 1,
+        "circle-stroke-width": 2,
+        "circle-radius": 5,
+      },
+      layout: {
+        visibility: "visible",
+      },
+    });
+
+    map.on("click", source.id + '.fill', handleLayerClick);
+    map.on("click", source.id + '.line', handleLayerClick);
+    map.on("click", source.id + '.circle', handleLayerClick);
+  }
+
+  function createRasterFeatureMapLayers(tileSource: Source, boundsSource: Source) {
+    const map = getMap();
+    const sourceIdentifiers = tileSource.id.split('.');
+    const metadata = {
+      layer_id: sourceIdentifiers[0],
+      layer_copy_id: sourceIdentifiers[1],
+      frame_id: sourceIdentifiers[2],
+    }
+
+    // Tile Layer
+    map.addLayer({
+      id: tileSource.id + '.tile',
+      type: "raster",
+      source: tileSource.id,
+      metadata,
+    });
+
+    map.addLayer({
+      id: boundsSource.id + '.mask',
+      type: "fill",
+      source: boundsSource.id,
+      paint: {
+        "fill-opacity": 0,
+      },
+    });
+
+    map.on("click", boundsSource.id + '.mask', handleLayerClick);
+  }
+
+  async function cacheRasterData(raster: RasterData) {
+    if (rasterTooltipDataCache.value[raster.id] !== undefined) {
+      return;
+    }
+    const data = await getRasterDataValues(raster.id);
+    rasterTooltipDataCache.value[raster.id] = data;
+  }
+
+  function createVectorTileSource(vector: VectorData, sourceId: string): Source | undefined {
+    const map = getMap();
+    const vectorSourceId = sourceId + '.vector.' + vector.id
+    map.addSource(vectorSourceId, {
+      type: "vector",
+      tiles: [`${baseURL}vectors/${vector.id}/tiles/{z}/{x}/{y}/`],
+    });
+    const source = map.getSource(vectorSourceId);
+    if (source) {
+      createVectorFeatureMapLayers(source);
+      return source;
+    }
+  }
+
+  function createRasterTileSource(raster: RasterData, sourceId: string): Source | undefined {
+    const map = getMap();
+
+    const params = {
+      projection: 'EPSG:3857'
+    }
+    const tileParamString = new URLSearchParams(params).toString();
+    const tilesSourceId = sourceId + '.raster.' + raster.id;
+    map.addSource(tilesSourceId, {
+      type: "raster",
+      tiles: [`${baseURL}rasters/${raster.id}/tiles/{z}/{x}/{y}.png/?${tileParamString}`],
+    });
+    const tileSource = map.getSource(tilesSourceId);
+
+    const bounds = raster.metadata.bounds;
+    const boundsSourceId = sourceId + '.bounds.' + raster.id;
+    let { xmin, xmax, ymin, ymax, srs } = bounds;
+    [xmin, ymin] = proj4(srs, "EPSG:4326", [xmin, ymin]);
+    [xmax, ymax] = proj4(srs, "EPSG:4326", [xmax, ymax]);
+    map.addSource(boundsSourceId, {
+      type: "geojson",
+      data: {
+        type: "Polygon",
+        coordinates: [[
+          [xmin, ymin], [xmin, ymax], [xmax, ymax], [xmax, ymin], [xmin, ymin],
+        ]]
+      }
+    });
+    const boundsSource = map.getSource(boundsSourceId);
+
+    if (tileSource && boundsSource) {
+      createRasterFeatureMapLayers(tileSource, boundsSource);
+      cacheRasterData(raster);
+      return tileSource;
+    }
+  }
+
+  function addLayerFrameToMap(frame: LayerFrame, sourceId: string) {
+    if (!mapSources.value[sourceId]) {
+      if (frame.vector) {
+        const vector = createVectorTileSource(frame.vector, sourceId);
+        if (vector) mapSources.value[sourceId] = vector;
+      }
+      if (frame.raster) {
+        const raster = createRasterTileSource(frame.raster, sourceId);
+        if (raster) mapSources.value[sourceId] = raster;
+      }
+    }
+  }
+
   return {
     // Data
     map,
@@ -115,6 +296,7 @@ export const useMapStore = defineStore('map', () => {
     showMapBaseLayer,
     tooltipOverlay,
     clickedFeature,
+    rasterTooltipDataCache,
     // Functions
     handleLayerClick,
     toggleBaseLayer,
@@ -123,5 +305,11 @@ export const useMapStore = defineStore('map', () => {
     getTooltip,
     setMapCenter,
     clearMapLayers,
+    createVectorFeatureMapLayers,
+    createRasterFeatureMapLayers,
+    createVectorTileSource,
+    createRasterTileSource,
+    addLayerFrameToMap,
+    cacheRasterData,
   }
 });
